AddStatement 클래스를 로컬 클래스로 이동
