+++ src/main/java/UserDao.java
@@ -22,6 +22,22 @@ public class UserDao {
     }
 
     public void add(User user) throws ClassNotFoundException, SQLException {
+        class AddStatement implements StatementStrategy {
+            private User user;
+            AddStatement(User user) {
+                this.user = user;
+            }
+
+            @Override
+            public PreparedStatement makePreparedStatement(Connection c) throws SQLException {
+                final PreparedStatement ps = c.prepareStatement("insert into users(id, name, password) values(?,?,?)");
+                ps.setString(1, user.getId());
+                ps.setString(2, user.getName());
+                ps.setString(3, user.getPassword());
+                return ps;
+            }
+        }
+
         final StatementStrategy strategy = new AddStatement(user);
         jdbcContextWithStatementStrategy(strategy);
     }
